Moved each page jsx and css file to their own folder in pages
